Move to using Heroku for frontend deployment
