--- client/now.json
@@ -1,11 +1,0 @@
-{
-    "version": 2,
-    "name": "knowledge",
-    "alias": "knowledge",
-    "builds": [
-        {
-			"src": "package.json",
-			"use": "@now/next"
-		}
-    ]
-}
