+++ client/package.json
@@ -9,7 +9,6 @@
 		"dev": "next",
 		"start": "next start",
 		"build": "next build",
-		"deploy": "now",
 		"test": "xo"
 	},
 	"dependencies": {
@@ -31,7 +30,6 @@
 		"eslint": "5.12.1",
 		"eslint-config-xo-react": "^0.18.0",
 		"eslint-plugin-react": "^7.12.4",
-		"now": "^13.1.2",
 		"xo": "^0.24.0"
 	},
 	"xo": {
