+++ client/readme.md
@@ -6,8 +6,6 @@ In the project directory, you can run:
 
 ### `npm start`
 
-### `npm build`
-
-### `npm run lint`
+### `npm run build`
 
 ### `npm test`
